service/requirements.txt
fastapi
uvicorn
httpx
motor
aiofiles
wheel
python-multipart
python-jose
passlib[bcrypt]